+++ client/src/components/Workflow/Editor/Tools/ToolBar.vue
